Change `href` to `url`
`badge.twig` and `node.html.twig` use `url` as the variable name.  There
are no other templates that use `href` as a variable, from what I can
see.

Added `render` filter to `url`, as in `badge.twig`.  Not sure
why it's there, but it doesn't appear to be causing problems.  If it
shouldn't be there, it can be removed from both places in the future.

Also adds parameter to docblock in response to #8, point 4:
https://www.drupal.org/project/radix/issues/3052990#comment-13101254